# docs/swagger.yaml
basePath: /
definitions:
  domain.FollowerDTO:
    properties:
      userId:
        type: integer
      userName:
        type: string
    type: object
  domain.Post:
    properties:
      category:
        type: integer
      date:
        type: string
      id_post:
        type: integer
      price:
        type: number
      product:
        $ref: '#/definitions/domain.Product'
      user_id:
        type: integer
    required:
    - category
    - price
    - user_id
    type: object
  domain.Product:
    properties:
      brand:
        type: string
      color:
        type: string
      notes:
        type: string
      product_id:
        type: integer
      product_name:
        type: string
      type:
        type: string
    required:
    - brand
    - color
    - notes
    - product_name
    - type
    type: object
  domain.RequestPostCreate:
    properties:
      category:
        type: integer
      date:
        description: Formato "DD-MM-YYYY"
        type: string
      price:
        type: number
      product:
        $ref: '#/definitions/domain.Product'
      user_id:
        type: integer
    required:
    - category
    - date
    - price
    - product
    - user_id
    type: object
  domain.RequestUserCreate:
    properties:
      email:
        type: string
      name:
        type: string
      role:
        type: string
    required:
    - email
    - name
    - role
    type: object
  domain.RequestUserUpdate:
    properties:
      email:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
  domain.User:
    properties:
      date:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
  domain.UserFollowersListResponse:
    properties:
      followers:
        items:
          $ref: '#/definitions/domain.FollowerDTO'
        type: array
      userId:
        type: integer
      userName:
        type: string
    type: object
  domain.UserListResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API para o desafio Social Meli do bootcamp.
  title: Social Meli API
  version: "1.0"
paths:
  /products/publish:
    post:
      consumes:
      - application/json
      description: Cria uma postagem de produto para um vendedor específico
      parameters:
      - description: Dados da postagem
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/domain.RequestPostCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Post'
        "400":
          description: Erro de validação
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar uma nova postagem
      tags:
      - Posts
  /users:
    get:
      description: Retorna uma lista contendo todos os usuários cadastrados no sistema
      produces:
      - application/json
      responses:
        "200":
          description: Lista de usuários retornada com sucesso
          schema:
            items:
              $ref: '#/definitions/domain.UserListResponse'
            type: array
        "500":
          description: Erro interno ao buscar usuários no banco
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar todos os usuários
      tags:
      - Users
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do usuário
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.RequestUserCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Dados inválidos ou JSON malformado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro interno ao salvar no banco
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Criar um novo usuário
      tags:
      - Users
  /users/{userId}:
    delete:
      parameters:
      - description: ID do Usuário
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Mensagem de sucesso
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Usuário não existe para ser deletado
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Deletar um usuário
      tags:
      - Users
    get:
      parameters:
      - description: ID do Usuário
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: ID inválido (não é um número)
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Usuário não encontrado no banco
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Buscar usuário por ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Atualiza os dados (nome e email) de um usuário existente pelo seu
        ID
      parameters:
      - description: ID do usuário a ser atualizado
        in: path
        name: userId
        required: true
        type: integer
      - description: Novos dados do usuário
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.RequestUserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: Usuário atualizado com sucesso
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: ID inválido ou JSON malformado
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Usuário não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Atualizar um usuário
      tags:
      - Users
  /users/{userId}/follow/{userIdToFollow}:
    post:
      consumes:
      - application/json
      description: Permite que um usuário siga um vendedor específico
      parameters:
      - description: ID do usuário que vai seguir
        in: path
        name: userId
        required: true
        type: integer
      - description: ID do vendedor a ser seguido
        in: path
        name: userIdToFollow
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Usuário seguido com sucesso
          schema:
            type: string
        "400":
          description: Erro na requisição
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Seguir um vendedor
      tags:
      - Users
  /users/{userId}/followed/list:
    get:
      consumes:
      - application/json
      description: Retorna a lista de todos os vendedores seguidos por um determinado
        usuário
      parameters:
      - description: ID do vendedor
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserFollowersListResponse'
        "404":
          description: Vendedor não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro ao buscar vendedores seguidos
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar todos os vendedores seguidos
      tags:
      - Users
  /users/{userId}/followers/count:
    get:
      description: Retorna a quantidade total de seguidores de um usuário, DESDE QUE
        ele seja um vendedor (role='seller')
      parameters:
      - description: ID do vendedor
        in: path
        name: userId
        required: true
        type: integer
      responses:
        "400":
          description: ID inválido ou usuário não é vendedor
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Obter contagem de seguidores
      tags:
      - Users
  /users/{userId}/followers/list:
    get:
      consumes:
      - application/json
      description: Retorna a lista de todos os usuários que seguem um vendedor específico
      parameters:
      - description: ID do vendedor
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserFollowersListResponse'
        "404":
          description: Vendedor não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Listar seguidores de um vendedor
      tags:
      - Users
  /users/{userId}/unfollow/{userIdToUnfollow}:
    put:
      consumes:
      - application/json
      description: Permite que um usuário pare de seguir um vendedor específico
      parameters:
      - description: ID do usuário
        in: path
        name: userId
        required: true
        type: integer
      - description: ID do vendedor a deixar de seguir
        in: path
        name: userIdToUnfollow
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deixou de seguir com sucesso
          schema:
            type: string
      summary: Deixar de seguir um vendedor
      tags:
      - Users
swagger: "2.0"